promise structure for client communication; added WET test file from group 2
