+++ server/relationship/relationship.js
@@ -2,37 +2,47 @@ var express = require('express');
 var router = express.Router();
 var fs = require('fs');
 var path = require('path');
-const StanfordCoreNLPClient = require('../middelwares/StanfordCoreNLPClient');
+const StanfordCoreNLPClient = require('../relationship/StanfordCoreNLPClient');
 
-const client = new StanfordCoreNLPClient(undefined, 'openie, coref', {"openie.resolve_coref": "true"});
-
-var data = "";
-
-
-fs.readFile(path.join(__dirname, '../resources') + '/testInput.txt',  "utf-8", function read(err, data) {
-    if (err) {
-        throw err;
+const client = new StanfordCoreNLPClient(
+    undefined,
+    'openie, coref',
+    {
+        'openie.resolve_coref': 'true'
     }
+);
 
-    // Invoke the next step here however you like
-    console.log(data);   // Put all of the code here (not the best solution)
-    processFile(data);          // Or put the next step in a function and invoke it
-});
+var data = '';
 
-function processFile(phrase){
-    client.annotate(phrase)
-        .then(function(result){
-            data = result;
+getFileContent('chaconne.txt')
+    .catch(function (error) {
+        console.log('error: reading file');
+    }).then(function (data) {
+        console.log('success: reading file');
+        return client.annotate(data);
+    }).catch(function (error) {
+        console.log('error: coreNLP processing');
+    }).then(function (result) {
+        console.log('success: coreNLP processing');
+        data = result;
+    });
+
+
+function getFileContent(filename) {
+    return new Promise(function (resolve, reject) {
+        fs.readFile(path.join(__dirname, '../resources') + '/' + filename, 'utf-8', function read(err, data) {
+            if (err) {
+                reject(err);
+            } else {
+                resolve(data);
+            }
         });
+    });
 }
-
-
 
 /* GET home page. */
 router.get('/', function (req, res, next) {
     res.send(data);
 });
 
-
 module.exports = router;
-
