+++ server/resources/chaconne.txt
@@ -1,0 +1,9 @@
+Chaconne, Italian Ciaccona, solo instrumental piece that forms the fifth and final movement of the Partita No. 2 in D Minor, BWV 1004, by Johann Sebastian Bach. Written for solo violin, the Chaconne is one of the longest and most challenging entirely solo pieces ever composed for that instrument.
+
+Bach’s string compositions, including a half dozen partitas and sonatas for solo violin, were composed in the late 1710s and early 1720s, while Bach was employed at the court in Köthen, Germany. It was a period of great freedom and creativity for the composer.
+
+The Chaconne forms the longest movement of the piece by far, making up roughly half of the entire partita. It draws upon the Baroque dance form known as a chaconne, in which a basic theme stated at the opening is then restated in several variations. In Bach’s Chaconne, the basic theme is four measures long, short and simple enough to allow for 64 variations. From a stern and commanding mood at the beginning, Bach gradually increases the complexity of his theme, mixing in various compositional effects. Some twists upon the theme are spacious and grand; others flow nimbly. Fast runs and large interval skips are frequent, requiring much dexterity from the performer. Bach also calls forth changes in emotional intensity, as some variations are dominated by long notes and others by many, more urgent short notes. Bach builds up his work over 256 measures, finally restating the theme at the end with new, even stronger harmonies.
+
+A century and a half after Bach composed the piece, Johannes Brahms wrote:
+
+The Chaconne is for me one of the most wonderful, incomprehensible pieces of music. On a single staff, for a small instrument, the man writes a whole world of the deepest thoughts and the most powerful feelings. If I were to imagine how I might have made, conceived the piece, I know for certain that the overwhelming excitement and awe would have driven me mad.
